Qr Code with MVVM
